[*] check back

+++ linux/debian/scripts/compile_and_pack.sh
@@ -75,10 +75,10 @@ if [[ $ARCH_VERSION == "amd64" ]]; then
 	&& ${CMAKE_PATH}cmake -DCMAKE_BUILD_TYPE=Debug ../ && make -j$(nproc) && ${CMAKE_PATH}cpack && repack *.deb && mv -v *.deb ../packages/ && rm -r * || error=$?
 	
 
-	
+	if [[ $CI_COMMIT_REF_NAME =~ ^.*-rwd$ ]]; then
 	echo "==== Building with reldebuginfo"
 	${CMAKE_PATH}cmake -DCMAKE_BUILD_TYPE=RelWithDebInfo ../ && make -j$(nproc) && ${CMAKE_PATH}cpack && repack *.deb && mv -v *.deb ../packages/ && rm -r * || error=$?
-
+	fi
 
 	sed -ibak 's/#set(BUILD_WITH_GDB_DRIVER_PGSQL ON)/set(BUILD_WITH_GDB_DRIVER_PGSQL ON)/' ../CMakeLists.txt || error=$?
 	${CMAKE_PATH}cmake ../ && make -j$(nproc) && ${CMAKE_PATH}cpack && repack *.deb && mv -v *.deb ../packages/ && rm -r * \
